Check for mu-session-id, added test method for data
A mu-session-id is necessary for knowing in what graph the data needs to
be put, so this throws an error for now if it is missing. This data test
method is for putting some testdata in a specific test graph. This is to
simulate the data flowing in the normal organisation graph.