Simplification of the distribution logic.
The root cause for this changes was a bug in the calculation of the
diffs.

It seems the
```
-  // Make shallow copy of the starting data store for making comparisons
-  const originalDataStore = new N3.Store();
-  dataStore.forEach((q) => originalDataStore.addQuad(q));
```

didn't really make a sound base to compare

```
-  // Make comparisons to the original data to only get updated values
-  const { toRemoveStore, toInsertStore } = compareStores(
-    originalDataStore,
-    resultDataStore
-  );
```

Since the call

```
-  // "Execute" the changesets as if they where queries on the internal store
-  const resultDataStore = await executeChangesets(
-    dataStore,
-    effectiveChangesets
-  );
```

affected the `orignalDataStore`. The quads not being a deepCopy. (Bit
unexpected addmitingly)

Diving further, I think we don't really need such fine grained
approach.
We just want the data from the source graph being reflected in the
vendor graph. So brutally flushing and copying on every update of the
subject should be sufficient.
This simplifies stuff a lot.